+++ lib/cb/config.rb
@@ -86,7 +86,9 @@ module Cb
       @uri_resume_put                     ||= '/cbapi/resumes/:resume_hash'
       @uri_resume_delete                  ||= '/cbapi/resumes/:resume_hash'
       @uri_resume_list                    ||= '/v3/resume/list'
-      @uri_hire_insider                     ||= '/cbapi/jobreport/:application_did'
+      @uri_resume_language_codes          ||= '/v1/languagecodes'
+      @uri_country_codes                  ||= '/v1/countrycodes'
+      @uri_hire_insider                   ||= '/cbapi/jobreport/:application_did'
     end
 
     def set_attr_accessors