+++ lib/cb/models/implementations/resumes/language_code.rb
@@ -1,0 +1,16 @@
+module Cb
+  module Models
+    class LanguageCode < ApiResponseModel
+      attr_accessor :code,:name
+
+      def set_model_properties
+        @code = api_response['Code']
+        @name = api_response['Name']
+      end
+
+      def required_fields
+        ['Code', 'Name']
+      end
+    end
+  end
+end