+++ lib/cb/utils/response_map.rb
@@ -45,6 +45,9 @@ module Cb
             Cb::Requests::Resumes::Put => Cb::Responses::Resume,
             Cb::Requests::Resumes::Delete => Cb::Responses::Resume,
             Cb::Requests::Resumes::List => Cb::Responses::ResumeList,
+            Cb::Requests::Resumes::LanguageCodes => Cb::Responses::LanguageCodes,
+
+            Cb::Requests::DataLists::CountryCodes => Cb::Responses::CountryCodes,
 
             Cb::Requests::User::ChangePassword => Cb::Responses::User::ChangePassword,
             Cb::Requests::User::CheckExisting => Cb::Responses::User::CheckExisting,