+++ spec/cb/utils/response_map_spec.rb
@@ -86,18 +86,26 @@ module Cb
         expect(response_map.response_for(request_namespace::List)).to eq response_namespace::List
       end
 
-      it 'should map resume get to the correct response object' do
+      it 'maps resume to the correct response object' do
         request_namespace = Cb::Requests::Resumes
         response_namespace = Cb::Responses
 
-        expect(response_map .response_for(request_namespace::Get)).to eq response_namespace::Resume
+        expect(response_map.response_for(request_namespace::Get)).to eq response_namespace::Resume
+        expect(response_map.response_for(request_namespace::LanguageCodes)). to eq response_namespace::LanguageCodes
+      end
+
+      it 'maps to correct data list response object' do
+        request_namespace = Cb::Requests::DataLists
+        response_namespace = Cb::Responses
+
+        expect(response_map.response_for(request_namespace::CountryCodes)). to eq response_namespace::CountryCodes
       end
 
       it 'should map resume recommendations to the correct response object' do
         request_namespace = Cb::Requests::Recommendations
         response_namespace = Cb::Responses
 
-        expect(response_map .response_for(request_namespace::Resume)).to eq response_namespace::Recommendations
+        expect(response_map.response_for(request_namespace::Resume)).to eq response_namespace::Recommendations
       end
     end
   end