+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Version History
 ====
     * All Version bumps are required to update this file as well!!
 ----
+* 14.3.1 Fixed missing response mapping
+* 14.3.0 Added country codes api call
+* 14.2.0 Added resume language codes api call
 * 14.1.2 Added count limit to resume recommendations call
 * 14.1.1 Added case for Job model "pay" when setting model properties
          so job search result pay field would be added to response as well